refactored test_in_child_process and channel into its own crate
