+++ crates/common/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod channel;
+#[cfg(feature = "test_utils")]
+pub mod test_utils;
